Form deletion refactoring
